+++ packages/designer/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.56](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.55...@vtj/designer@0.8.56) (2024-06-22)
+
+
+### Features
+
+* ✨ api管理支持搜索，帮助中心支持返回首页 ([cbc1e9c](https://gitee.com/newgateway/vtj/commits/cbc1e9c0eb2fe15e52b98491178c2ba5ed56f423))
+
+
+
+
+
+## [0.8.55](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.54...@vtj/designer@0.8.55) (2024-06-21)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.54](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.53...@vtj/designer@0.8.54) (2024-06-20)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
 ## [0.8.53](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.52...@vtj/designer@0.8.53) (2024-06-20)
 
 **Note:** Version bump only for package @vtj/designer